shopify-store-chatgpt-app/pyproject.toml
[project]
name = "shopify-store-chatgpt-app"
version = "0.1.0"
description = "Shopify Store MCP Server with React UI Components for ChatGPT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "backoff>=2.2.1",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.3",
    "ShopifyAPI>=12.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "starlette>=0.41.0",
    "uvicorn>=0.35.0",
    "google-generativeai>=0.8.5",
]




